+++ src/settings/moduleSettings.ts
@@ -1,11 +1,11 @@
-type Dict = { [key: string]: unknown };
-
-class SettingStore {
-    public moduleSettings: Dict = {};
-}
-
-export const store = new SettingStore();
-
+import store, { SettingStore } from './settingsStore';
+/**
+ *
+ *
+ * @export
+ * @class ModuleSettings
+ * @template T
+ */
 export class ModuleSettings<T> {
     protected store: SettingStore;
     public key: string;
@@ -15,7 +15,12 @@ export class ModuleSettings<T> {
         this.key = '';
         this.setup<T>(key, initialState);
     }
-
+    /**
+     * Setter function fro setting the module settings
+     * TODO: Persist data in local storage
+     * @param {Partial<T>} value
+     * @memberof ModuleSettings
+     */
     set = (value: Partial<T>): void => {
         if (typeof this.store.moduleSettings[this.key] === 'object')
             this.store.moduleSettings[this.key] = {
@@ -24,20 +29,45 @@ export class ModuleSettings<T> {
             };
         else this.store.moduleSettings[this.key] = value;
     };
-
+    /**
+     * Getter function for retrieving the module settings registered by module by it's
+     * unique key.
+     *
+     * @memberof ModuleSettings
+     */
     get = (): T => {
         return this.store.moduleSettings[this.key] as T;
     };
-
+    /**
+     * Clear the settings object for this module.
+     * TODO: Persist in local storage
+     * @memberof ModuleSettings
+     */
     clear = (): void => {
         this.store.moduleSettings[this.key] = {};
     };
-
+    /**
+     * Function for checking unique keys in modules settings
+     *
+     * @private
+     * @param {string} key
+     * @return {*}  {boolean}
+     * @memberof ModuleSettings
+     */
     private isKeyUnique(key: string): boolean {
         if (key.length === 0) throw TypeError('Please enter a key!');
         return !this.store.moduleSettings.hasOwnProperty(key);
     }
-
+    /**
+     * Setup function use internal or in derived classes for this class.
+     * This will check if your key is Unique and trow TypeError if not.
+     *
+     * @protected
+     * @template T
+     * @param {string} key
+     * @param {Partial<T>} initialState
+     * @memberof ModuleSettings
+     */
     protected setup<T>(key: string, initialState: Partial<T>): void {
         if (this.isKeyUnique(key)) {
             this.store.moduleSettings[key] = initialState;
@@ -46,30 +76,4 @@ export class ModuleSettings<T> {
             throw TypeError('Key needs to be uniq!');
         }
     }
-}
-
-export class EchoModuleSettings<T> extends ModuleSettings<T> {
-    constructor(key: string, initialState: T, clear = false) {
-        super(key, initialState);
-        if (clear) {
-            this.clearStore();
-            this.setup(key, initialState);
-        }
-    }
-
-    private clearStore(): void {
-        Object.keys(this.store.moduleSettings).forEach((key) => delete this.store.moduleSettings[key]);
-    }
-
-    getAll(): unknown {
-        return this.store.moduleSettings;
-    }
-
-    getByKey = (key: string): Partial<T> | undefined => {
-        const index = Object.keys(this.store.moduleSettings).findIndex((objectKey: string) => objectKey === key);
-        if (index > -1) {
-            return this.store.moduleSettings[key] as T;
-        }
-        return undefined;
-    };
 }
